Split git adapter content into dedicated files
